move internal rest endpoints into separate package for easy reusability
